Задание к уроку "Интеграционное тестирование" (#3)
* User can create question

* User can view the list of questions

* User can view the question and list of answers for this

* User can answer the question

* User can register

* User can sign in

* User can logout

* Only authenticated user can create question

* Only authenticated user can answer the question

* Only authenticated author can delete own question

* Only authenticated author can delete own answer

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,21 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, only: %i[new create destroy]
+
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { question.answers.new }
+
+  # This is a stub, used for indexing in before_action :authenticate_user!
+  # Redefined in Decent Exposure
+  def new; end
 
   def create
+    question.user = current_user
+
     if question.save
-      redirect_to question
+      redirect_to question, notice: t('.success')
     else
       render :new
     end
@@ -21,9 +30,13 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user&.author?(question)
+      question.destroy
 
-    redirect_to questions_path
+      redirect_to questions_path, notice: t('.success')
+    else
+      render :show
+    end
   end
 
   private
